Move many SpecimenManager methods to SpecimenManagerNew in study-api. Move CohortFilter to study-api.
