--- specimen/src/org/labkey/specimen/SpecimenContainerListener.java
@@ -1,55 +1,0 @@
-/*
- * Copyright (c) 2021 LabKey Corporation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.labkey.specimen;
-
-import org.jetbrains.annotations.NotNull;
-import org.labkey.api.data.Container;
-import org.labkey.api.data.ContainerManager.ContainerListener;
-import org.labkey.api.security.User;
-import java.util.Collections;
-import java.util.Collection;
-
-import java.beans.PropertyChangeEvent;
-
-public class SpecimenContainerListener implements ContainerListener
-{
-    @Override
-    public void containerCreated(Container c, User user)
-    {
-    }
-
-    @Override
-    public void containerDeleted(Container c, User user)
-    {
-    }
-
-    @Override
-    public void propertyChange(PropertyChangeEvent evt)
-    {
-    }
-
-    @Override
-    public void containerMoved(Container c, Container oldParent, User user)
-    {
-    }
-
-    @NotNull @Override
-    public Collection<String> canMove(Container c, Container newParent, User user)
-    {
-        return Collections.emptyList();
-    }
-}
