+++ study/src/org/labkey/study/CohortFilterFactory.java
@@ -21,12 +21,13 @@ import org.labkey.api.data.Container;
 import org.labkey.api.data.ConvertHelper;
 import org.labkey.api.query.FieldKey;
 import org.labkey.api.security.User;
+import org.labkey.api.study.CohortFilter;
+import org.labkey.api.study.CohortFilter.Type;
 import org.labkey.api.study.Study;
 import org.labkey.api.study.StudyService;
 import org.labkey.api.util.Pair;
 import org.labkey.api.util.SafeToRenderEnum;
 import org.labkey.api.view.ActionURL;
-import org.labkey.study.CohortFilter.Type;
 import org.labkey.study.model.CohortImpl;
 import org.labkey.study.model.StudyManager;
