+++ study/src/org/labkey/study/specimen/report/SpecimenVisitReport.java
@@ -25,13 +25,13 @@ import org.labkey.api.security.User;
 import org.labkey.api.specimen.SpecimenTypeLevel;
 import org.labkey.api.specimen.report.SpecimenReportCellData;
 import org.labkey.api.specimen.report.SpecimenReportTitle;
+import org.labkey.api.study.CohortFilter;
 import org.labkey.api.study.Study;
 import org.labkey.api.study.StudyService;
 import org.labkey.api.study.model.ParticipantGroup;
 import org.labkey.api.util.PageFlowUtil;
 import org.labkey.api.view.ActionURL;
 import org.labkey.api.view.NotFoundException;
-import org.labkey.study.CohortFilter;
 import org.labkey.study.model.ParticipantGroupManager;
 import org.labkey.study.model.StudyManager;
 import org.labkey.study.model.VisitImpl;
