+++ study/src/org/labkey/study/specimen/report/participant/ParticipantVisitReport.java
@@ -16,16 +16,16 @@
 package org.labkey.study.specimen.report.participant;
 
 import org.labkey.api.data.SimpleFilter;
+import org.labkey.api.specimen.report.SpecimenReportTitle;
+import org.labkey.api.specimen.report.SummaryByVisitParticipant;
+import org.labkey.api.study.CohortFilter;
 import org.labkey.api.study.StudyService;
 import org.labkey.api.util.DemoMode;
 import org.labkey.api.view.ActionURL;
-import org.labkey.study.CohortFilter;
 import org.labkey.study.SpecimenManager;
-import org.labkey.api.specimen.report.SummaryByVisitParticipant;
 import org.labkey.study.controllers.specimen.SpecimenController;
 import org.labkey.study.model.StudyManager;
 import org.labkey.study.model.VisitImpl;
-import org.labkey.api.specimen.report.SpecimenReportTitle;
 import org.labkey.study.specimen.report.SpecimenVisitReport;
 import org.labkey.study.specimen.report.SpecimenVisitReportParameters;
