+++ study/src/org/labkey/study/visitmanager/AbsoluteDateVisitManager.java
@@ -21,8 +21,8 @@ import org.jetbrains.annotations.Nullable;
 import org.labkey.api.data.SQLFragment;
 import org.labkey.api.data.TableInfo;
 import org.labkey.api.security.User;
+import org.labkey.api.study.CohortFilter;
 import org.labkey.api.study.DataspaceContainerFilter;
-import org.labkey.study.CohortFilter;
 import org.labkey.study.StudySchema;
 import org.labkey.study.model.QCStateSet;
 import org.labkey.study.model.StudyImpl;
