+++ study/src/org/labkey/study/visitmanager/RelativeDateVisitManager.java
@@ -29,10 +29,10 @@ import org.labkey.api.query.DefaultSchema;
 import org.labkey.api.query.FieldKey;
 import org.labkey.api.query.FilteredTable;
 import org.labkey.api.security.User;
+import org.labkey.api.study.CohortFilter;
 import org.labkey.api.study.DataspaceContainerFilter;
 import org.labkey.api.study.Study;
 import org.labkey.api.study.model.ParticipantGroup;
-import org.labkey.study.CohortFilter;
 import org.labkey.study.StudySchema;
 import org.labkey.study.model.QCStateSet;
 import org.labkey.study.model.StudyImpl;
